moved Stop Conditions to a separate file
